+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [unreleased]
+
+### Deprecated
+
+- `ui/MarqueeDecorator` config `className` to be removed in 5.0.0
+
+### Added
+
+- `ui/MarqueeDecorator` config `css` to support customizing the MarqueeBase styles
+
 ## [4.5.0-alpha.1] - 2022-04-15
 
 ### Fixed